+++ test/lib/parse/query/constraints/greater_than_test.rb
@@ -5,7 +5,7 @@ class TestGreaterThanConstraint < Minitest::Test
   include ConstraintTests
 
   def setup
-    @klass = Parse::GreaterThanConstraint
+    @klass = Parse::Constraint::GreaterThanConstraint
     @key = :$gt
     @operand = :gt
     @keys = [:gt, :greater_than, :after]